Provide test to showcase the problems with Jackson's JsonSubTypes.
The MongoEventTest tries to insert and fetch an instance of the polymorphic `Event` model.
The test fails when executed against an actual MongoDB instance with the following error message:
`Caused by: com.fasterxml.jackson.databind.JsonMappingException: class com.fasterxml.jackson.core.util.JsonParserSequence cannot be cast to class org.immutables.criteria.mongo.bson4jackson.Wrapper`

To run the test against an actual MongoDB instance, set the system property `-Dmongo=mongodb://localhost` as described in the `MongoExtension`.

The exception that occurs comes from the false assumption in the `JacksonCodecs` that the `JsonParser` can be cast into a `Wrapper<BsonReader>`.
That assumption is false as soon as `@JsonSubTypes` is used, because the `BsonParser` gets wrapped by a `JsonParserSequence` and can no longer be cast into a `Wrapper<BsonReader>`.